+++ packages/vue/src/index.ts
@@ -1,8 +1,15 @@
-export { RobotGrid } from "./RobotGrid";
-export { RobotGrid as MachTable } from "./RobotGrid";
+import { RobotGrid } from "./RobotGrid";
+import { MachTablePlugin } from "./plugin";
+
+export { RobotGrid };
+export { RobotGrid as MachTable };
 export type { RobotGridVueProps } from "./RobotGrid";
 export { vueCellRenderer, vueDetailRenderer } from "./adapters";
 export { useMachTable } from "./useMachTable";
 export type { UseMachTableReturn } from "./useMachTable";
 export type { VueCellRendererProps } from "./adapters";
+export { MachTablePlugin } from "./plugin";
+export type { MachTablePluginOptions } from "./plugin";
 export * from "@agile-team/mach-table";
+
+export default MachTablePlugin;