+++ packages/vue/tsup.config.ts
@@ -1,8 +1,12 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: {
+    index: "src/index.ts",
+    async: "src/async.ts"
+  },
   format: ["esm", "cjs"],
+  splitting: true,
   dts: true,
   sourcemap: true,
   clean: true,